Add site with upstream SDC content, unedited
This pulls in almost all of its content from
https://github.com/CodeYourFuture/curriculum/tree/main/common-content
but individual content can be overridden as needed.

There are a few TODOs to address, and we will also want to modify e.g.
day plans to split into different days, but we will pick those up
separately.

But this shows the shape of how the site will inherit rather than fork
content.

This was pilled in as of upstream commit
83e619c7f961e9acd574b9e57561984a6fa763d1 on 2026-03-16 (as evidenced in
the go.mod file).